requirements.txt
asgiref==3.5.0
Django==4.0.1
django-widget-tweaks==1.4.12
sqlparse==0.4.2
tzdata==2021.5

Pillow~=9.0.1